Fixed error in using Wire.write(0x00) (it's now Wire.write((uint8_t) 0x00)).

+++ RTClib.cpp
@@ -132,7 +132,7 @@ uint8_t RTC_DS1307::begin(void) {
 
 uint8_t RTC_DS1307::isrunning(void) {
   Wire.beginTransmission(DS1307_ADDRESS);
-  Wire.write(0x00);	
+  Wire.write((uint8_t)0x00);	
   Wire.endTransmission();
 
   Wire.requestFrom(DS1307_ADDRESS, 1);
@@ -142,7 +142,7 @@ uint8_t RTC_DS1307::isrunning(void) {
 
 void RTC_DS1307::adjust(const DateTime& dt) {
     Wire.beginTransmission(DS1307_ADDRESS);
-    Wire.write(0x00);
+    Wire.write((uint8_t)0x00);	
     Wire.write(bin2bcd(dt.second()));
     Wire.write(bin2bcd(dt.minute()));
     Wire.write(bin2bcd(dt.hour()));
@@ -150,13 +150,13 @@ void RTC_DS1307::adjust(const DateTime& dt) {
     Wire.write(bin2bcd(dt.day()));
     Wire.write(bin2bcd(dt.month()));
     Wire.write(bin2bcd(dt.year() - 2000));
-    Wire.write(0x00);
+    Wire.write((uint8_t)0x00);	
     Wire.endTransmission();
 }
 
 DateTime RTC_DS1307::now() {
   Wire.beginTransmission(DS1307_ADDRESS);
-  Wire.write(0x00);
+    Wire.write((uint8_t)0x00);	
   Wire.endTransmission();
   
   Wire.requestFrom(DS1307_ADDRESS, 7);
